+++ gw2buildutil/defnfile/__init__.py
@@ -2,30 +2,30 @@ import io
 import re
 
 from .. import api, build
-from . import util, text as parse_text, section
+from . import parseutil, section
 
 _SKIP_SECTION = object()
 
 
 title_pattern = re.compile('^'
-    f'(?P<mode>{parse_text.words_pattern}) '
-    f'(?P<prof>{parse_text.word_pattern}) '
+    f'(?P<mode>{parseutil.words_pattern}) '
+    f'(?P<prof>{parseutil.word_pattern}) '
     '\\((?P<labels>'
-         f'{parse_text.sep_pattern(", ", parse_text.words_pattern)}'
+         f'{parseutil.sep_pattern(", ", parseutil.words_pattern)}'
          ')\\)'
     '$')
 
 def parse_title (title, api_storage):
     match = title_pattern.match(title)
     if match is None:
-        raise util.ParseError('title doesn\'t match expected format: '
-                              '{}'.format(repr(title)))
+        raise parseutil.ParseError('title doesn\'t match expected format: '
+                                   '{}'.format(repr(title)))
     fields = match.groupdict()
 
     game_mode = build.GameModes.from_id(fields['mode'])
     if game_mode is None:
-        raise util.ParseError('title doesn\'t start with a known '
-                              'game modes identifier: {}'.format(repr(title)))
+        raise parseutil.ParseError('title doesn\'t start with a known '
+                                   f'game modes identifier: {repr(title)}')
 
     try:
         profession = api_storage.from_id(api.entity.Profession, fields['prof'])
@@ -36,8 +36,8 @@ def parse_title (title, api_storage):
         except KeyError:
             elite_spec = None
         if elite_spec is None or not elite_spec.is_elite:
-            raise util.ParseError('title has a missing or incorrect profession '
-                                  f'identifier: {repr(title)}')
+            raise parseutil.ParseError('title has a missing or incorrect '
+                                       f'profession identifier: {repr(title)}')
         else:
             profession = elite_spec.profession
     else:
@@ -63,10 +63,11 @@ def parse_body (f, meta, api_storage):
              for line in f)
     build_data = {}
 
-    for title, section_lines in util.split_sections(lines, 'intro'):
+    for title, section_lines in parseutil.split_sections(lines, 'intro'):
         section_module = section_parsers.get(title)
         if section_module is None:
-            raise util.ParseError('unknown section: {}'.format(repr(title)))
+            raise parseutil.ParseError(
+                'unknown section: {}'.format(repr(title)))
         elif section_module is _SKIP_SECTION:
             for line in section_lines:
                 pass