+++ gw2buildutil/defnfile/section/encounters.py
@@ -1,4 +1,4 @@
-from .. import util
+from ... import util
 
 
 def parse (lines, meta, api_storage):